+++ src/base.js
@@ -1,6 +1,8 @@
 import * as THREE from 'three';
 import fontDefinitionHelvetikerBold from 'three/examples/fonts/helvetiker_bold.typeface.json';
-import { getRoundedRectShape, subtractMesh, getBoundingBoxSize } from './utils';
+import {
+  getRoundedRectShape, getCustomRoundedRectShape, subtractMesh, unionMesh, getBoundingBoxSize,
+} from './utils';
 
 class BaseTag3D {
   constructor(options) {
@@ -35,6 +37,7 @@ class BaseTag3D {
     this.baseMesh = null;
     this.borderMesh = null;
     this.subtitleMesh = null;
+    this.keychainAttachmentMesh = null;
     this.combinedMesh = null;
     this.exportedMeshes = [];
   }
@@ -148,6 +151,79 @@ class BaseTag3D {
     return borderMesh;
   }
 
+  /**
+   * @return {THREE.Mesh} the mesh of the keychain attachment hole
+   */
+  getKeychainAttachmentMesh() {
+    const holeRadius = 3;
+    const cornerPlacementOffset = 3;
+    const height = 9;
+    const width = height + cornerPlacementOffset;
+
+    const attachmentShape = getCustomRoundedRectShape(
+      -height / 2,
+      -width / 2,
+      height,
+      width,
+      0,
+      0,
+      height / 2,
+      height / 2,
+    );
+
+    const attachmentShapeMesh = new THREE.Mesh(new THREE.ExtrudeGeometry(attachmentShape, {
+      steps: 1,
+      depth: this.options.base.depth,
+      bevelEnabled: false,
+    }), this.materialBase);
+    attachmentShapeMesh.position.z = 0;
+    attachmentShapeMesh.updateMatrix();
+
+    const holeMesh = new THREE.Mesh(new THREE.CylinderGeometry(holeRadius, holeRadius, this.options.base.depth, 32), this.materialBase);
+    holeMesh.rotation.x = -Math.PI / 2;
+    holeMesh.position.z = this.options.base.depth / 2;
+    holeMesh.position.y = -cornerPlacementOffset / 2;
+    holeMesh.updateMatrix();
+
+    let finalMesh = subtractMesh(attachmentShapeMesh, holeMesh);
+
+    if (this.options.base.keychainPlacement === 'left') {
+      finalMesh.position.x = 0;
+      finalMesh.position.y = -this.options.base.width / 2 - width / 2 + cornerPlacementOffset;
+      finalMesh.position.z = 0;
+    } else if (this.options.base.keychainPlacement === 'top') {
+      finalMesh.position.x = -this.options.base.height / 2 - height / 2 + cornerPlacementOffset / 2;
+      finalMesh.position.y = 0;
+      finalMesh.position.z = 0;
+      finalMesh.rotation.z = -Math.PI / 2;
+    } else if (this.options.base.keychainPlacement === 'topLeft') {
+      finalMesh.position.x = -this.options.base.height / 2;
+      finalMesh.position.y = -this.options.base.width / 2;
+      finalMesh.position.z = 0;
+      finalMesh.rotation.z = -Math.PI / 4;
+    }
+    finalMesh.updateMatrix();
+
+    if (this.options.base.mirrorHoles) {
+      const mirror = subtractMesh(attachmentShapeMesh, holeMesh);
+      if (this.options.base.keychainPlacement === 'left') {
+        mirror.position.y = this.options.base.width / 2 + width / 2 - cornerPlacementOffset;
+        mirror.rotation.z = Math.PI;
+      } else if (this.options.base.keychainPlacement === 'top') {
+        mirror.position.x = this.options.base.height / 2 + height / 2 - cornerPlacementOffset / 2;
+        mirror.rotation.z = Math.PI / 2;
+      } else if (this.options.base.keychainPlacement === 'topLeft') {
+        mirror.position.x = this.options.base.height / 2;
+        mirror.position.y = this.options.base.width / 2;
+        mirror.rotation.z = -Math.PI / 4 + Math.PI;
+      }
+      mirror.updateMatrix();
+      finalMesh = unionMesh(finalMesh, mirror);
+    }
+
+    return finalMesh;
+  }
+
   getCornerRadius() {
     if (this.options.base.shape === 'roundedRectangle') {
       return this.options.base.cornerRadius;
@@ -191,6 +267,10 @@ class BaseTag3D {
       combinedGeometry.merge(this.subtitleMesh.geometry, this.subtitleMesh.matrix);
     }
 
+    if (this.keychainAttachmentMesh) {
+      combinedGeometry.merge(this.keychainAttachmentMesh.geometry, this.keychainAttachmentMesh.matrix);
+    }
+
     const combinedMesh = new THREE.Mesh(combinedGeometry, this.materialBase);
     return combinedMesh;
   }
@@ -212,6 +292,11 @@ class BaseTag3D {
       if (!this.options.code.invert) {
         this.exportedMeshes.push(this.subtitleMesh);
       }
+    }
+
+    if (this.options.base.hasKeychainAttachment) {
+      this.keychainAttachmentMesh = this.getKeychainAttachmentMesh();
+      this.exportedMeshes.push(this.keychainAttachmentMesh);
     }
   }
 }