+++ src/components/SpotifyMenu.vue
@@ -122,6 +122,9 @@ export default {
           textSize: 8,
           textMessage: '',
           textDepth: 1,
+          hasKeychainAttachment: false,
+          keychainPlacement: 'left',
+          mirrorHoles: false,
         },
         code: {
           depth: 1,